framework/requirements.txt
psutil
termcolor
prettytable
pyshark
python-gitlab
pyyaml
charset-normalizer~=2.0.0
Jinja2 ==3.1.4
click ==8.1.3
colorama ==0.4.4
cryptography ==42.0.4
python-dateutil ==2.8.2
python-docs-theme ==2024.4
requests ==2.32.3
sphinx-click ==3.0.2
sphinx ==7.3.7
sphinxcontrib-applehelp <1.0.8
sphinxcontrib-devhelp <1.0.6
sphinxcontrib-htmlhelp <2.0.5
sphinxcontrib-qthelp <1.0.7
sphinxcontrib-serializinghtml <1.1.10
tqdm ==4.66.3
wheel
